Created the real-time data portion of the project that takes real-time stock data from Google server and every 10 seconds takes data from the server, calculates all the attributes from the prices for each stock and will have it eventually be fed to the new recurrent neural network portion
Several new changes have been made to the project. The first and
most important is that rather than having real-time learning algorithm
work on the data, I've chosen to utilize a recurrent neural network
algorithm. The reason for this change is due to the fact that I
was not getting the results I wanted, and it didn't have the complexity
needed to work well in it's prediction. I chose to try a recurrent
neural network instead for a few reasons.

First, it allows for much more complexity, in that it can take a lot
more attributes as input without the extreme cost in time and it also
allowed to utilize and work on an algorithm that is cutting edge in the
AI industry and is being utilized heavily in real-world environments. It
allows me to get practice with deep-learning while being able to solve
wildly complex problems like stock-market prediction much better than
reinforcement learning could have.

The new changes are:
-Retrieving real-time data from a google server that processes stock data
 in real-time. This is crucial in allowing  for the most chance at
 success.
-Retrieving stock split and dividend information from a website by
 scraping the website for which companies have upcoming stock splits
 or dividends upcoming and when.
-A backup option for retrieving real-time stock information due to the
 fact that occassionally google detects your data retrieval as a bot and
 temporarily bans you. If this happens, the program utilizes yahoo finance
 and scrapes the information from there website. This is only a backup due
 to the fact that the timing takes roughly 25 times as long, due to the
 need to get info from individual url pages instead of all at once from a
 server.
-A function to find the total stock movement in the market. This is done
 by testing whether each stock is up or down in the day and subtracting
 the up and down movement from one another to form one integer
 representing the market for the day. It does this calculation every
 5 minutes due to the heavy processing needed to test each stock in both
 the nasdaq and nyse rather than our normal list of about 50 stocks we're
 following.
-The program is setup to run everyday at the start of the market and end
 at the end of the market day. After which is makes some ending changes
 to update the closing day values.
-Also added a function that will take the stock news at the begginning of
 the stock data for each stock we're following using google's server. It
 returns a list of articles dealing with that particle stock for that day
 if any. This function is not being utilized currently but will eventually
 be used to assist in prediction.