Created get_neural_inout() which takes as parameters a list w/ two items, the first is a string with the chosen stock ticker you wanted to predict on as a string. The second was an integer representing the period length you wanted to predict for.
This function would open two files using the ticker, the first
was the neural network that had been trained already, as well as
the list of indicators that was used to train the network and will
be used to predict on the network as well.

We then use the list of indicator template and change the default
predicting stock(which is AAPL) and change it to our chosen name
as well as change each return indicator for the base etfs/indexes
to be our period length return indicator.

Once this is done, we open our base etfs/indexes as well as our
chosen company dataframes containing all of our historical indicators.
We then combine the columns we're using for each etf/index/stock so we're
left w/ the indicators used to train the network. Then we return only the
latest 5 indicators, and latest 4 outputs.

We then return the combined dataframe, the output
dataframe, the column