Added 2 seperate files, a indicator list and a trained network
For these additions, the first was an indicator list that you
could use as your practice training inputs. These are the names
of the indicators you would use from the indicator dataframes. The
second file was a trained network so you could test on an already
trained network to see how it works. It uses the same indicators as
the first files indicators.