Added the Capstone project technical analysis paper
This capstone paper is for my final project for my Udacity course
that is used for explaining the intracies of this project, why it
was made, and exactly how to use it. It also goes into great detail
of the process of creating this application, and who exactly this
is meant for. It goes in to what technologies were used in this
application and how they were utilized.