Combined nyse_dates_prds.py and calc_ind.py into single file and updated references to them in RealTimeCalculator and RealTimeNeuralNetork
Combined the two python files nyse_dates_prds and calc_ind into
a single file using the nyse_dates_prds name. This change was done
because calc_ind was using the same dating function from the nyse
file and the other two functions were small additions that could
easily be added. This allowed for adjusting the file references in
the RealTimeCalculator and RealTimeNeuralNetwork ipython files to
use only one file instead of two.

The other change made was adding the dating function from the
neural network ipython file to the nyse_dates_prds python file
due to the similarities to the other two dating files with just
different default parameters.