Added comments for the create_indicators function in RealTimeNeuralNetwork notebook
Added comments to the create_indicators function which is utilized if
you don't already have an indicator list for the stock your predicting
for. If it's called it will create the indicator list for training and
testing and the comments were added to help the reader understand whats
happening and how to change it if necessary.