Switched from using find() to a regex solution
In the backup_yahoo function in the realtimecalculator notebook, I
switched from using the python find function to using a combination
of beautifulsoup to find a specific object that contained the
price of the stock and a regular expression to take that object and
grab the price from it. This was switched because of how easy it
was for the site to switch something and mess up the function but
this function is now more solid to changes.