Converted split/dividend funcs to use regular expressions
Changed the split and dividend functions in the realtimecalc notebook
due to the need for a more robust method to changes. I changed from
using simple find methods to find the info I need, to using beautiful
soup objects to get the table data and then using regular expressions
to get the exact data and remove anything I don't need. This is done
so that if the website changes its design at any point, I won't need
to change my function due to its beautiful soup method of getting
data.